End of section 18

+++ client/src/app/features/checkout/checkout.component.ts
@@ -16,6 +16,8 @@ import { CartService } from '../../core/services/cart.service';
 import { CurrencyPipe } from '@angular/common';
 import { Router } from '@angular/router';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { OrderToCreate, ShippingAddress } from '../../shared/models/order';
+import { OrderService } from '../../core/services/order.service';
 
 @Component({
   selector: 'app-checkout',
@@ -29,6 +31,7 @@ export class CheckoutComponent implements OnInit, OnDestroy {
   private stripeService = inject(StripeService);
   private accountService = inject(AccountService);
   public cartService = inject(CartService);
+  public orderService = inject(OrderService);
   public router = inject(Router);
   private snackbar = inject(SnackbarService);
   addressElement?: StripeAddressElement;
@@ -116,16 +119,16 @@ export class CheckoutComponent implements OnInit, OnDestroy {
         const result = await this.stripeService.confirmPayment(this.confirmationToken);
 
         if (result.paymentIntent?.status === 'succeeded') {
-          //const order = await this.createOrderModel();
-          //const orderResult = await firstValueFrom(this.orderService.createOrder(order));
-          //if (orderResult) {
-          //  this.orderService.orderComplete = true;
+          const order = await this.createOrderModel();
+          const orderResult = await firstValueFrom(this.orderService.createOrder(order));
+          if (orderResult) {
+            this.orderService.orderComplete = true;
             this.cartService.deleteCart();
             this.cartService.selectedDelivery.set(null);
             this.router.navigateByUrl('/checkout/success');
-          //} else {
-          //  throw new Error('Order creation failed');
-          //} 
+          } else {
+            throw new Error('Order creation failed');
+          } 
         } else if (result.error) {
           throw new Error(result.error.message);
         } else {
@@ -139,14 +142,36 @@ export class CheckoutComponent implements OnInit, OnDestroy {
       this.loading = false;
     }
   }
+
+  private async createOrderModel(): Promise<OrderToCreate> {
+    const cart = this.cartService.cart();
+    const shippingAddress = await this.getAddressFromStripeAddress() as ShippingAddress;
+    const card = this.confirmationToken?.payment_method_preview.card;
+
+    if (!cart?.id || !cart.deliveryMethodId || !card || !shippingAddress) {
+      throw new Error('Problem creating order');
+    }
+
+    return {
+      cartId: cart.id,
+      paymentSummary: {
+        last4: +card.last4,
+        brand: card.brand,
+        expMonth: card.exp_month,
+        expYear: card.exp_year
+      },
+      deliveryMethodId: cart.deliveryMethodId,
+      shippingAddress
+    }
+  }
   
-  private async getAddressFromStripeAddress(): Promise<Address | null> {
+  private async getAddressFromStripeAddress(): Promise<Address | ShippingAddress | null> {
     const result = await this.addressElement?.getValue();
     const address = result?.value.address;
 
     if (address) {
       return {
-        //name: result.value.name,
+        name: result.value.name,
         line1: address.line1,
         line2: address.line2 || undefined,
         city: address.city,
